requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-socketio==5.5.2
celery==5.2.3
redis==4.1.0
pandas==1.3.3
openpyxl==3.0.9
python-multipart==0.0.5
pydantic==1.8.2
python-dotenv==0.19.0